% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_ema_cache}
\alias{clear_ema_cache}
\title{Clear Local EMA Data Cache}
\usage{
clear_ema_cache(name = NULL)
}
\arguments{
\item{name}{Optional string. If provided (e.g., "event"), only deletes
that specific cache. If NULL (default), wipes the entire EMAdownload cache folder.}
}
\description{
Deletes the local .rds files and ETag logs stored in the
user's AppData/Cache folder.
}
