% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ema_catch.R
\name{get_ema_catch}
\alias{get_ema_catch}
\title{Get EMA catch table}
\usage{
get_ema_catch(force_download)
}
\arguments{
\item{force_download}{Bypass cache and force download}
}
\value{
a dataframe with all trawl related catch information
}
\description{
This function pulls ecosystem survey data (from EMA surveys) from the AKFIN data server
It includes parameters to control start and end year, gear type, and TSN (species code)
}
