% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ema_event.R
\name{get_ema_event}
\alias{get_ema_event}
\title{Get ema event table}
\usage{
get_ema_event(force_download = FALSE)
}
\arguments{
\item{force_download}{Bypass cache and force download}
}
\value{
dataframe with event infomration for most gear types from surveys
}
\description{
This function pulls ecosystem survey data (from EMA surveys) from the AKFIN data server
}
