% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ema_fish.R
\name{get_ema_fish}
\alias{get_ema_fish}
\title{Get EMA fish specimen data}
\usage{
get_ema_fish(force_download = FALSE)
}
\arguments{
\item{force_download}{Bypass cache and force download}
}
\value{
a dataframe with all specimen data
}
\description{
This function pulls specimen level data from
ecosystem survey data (from EMA surveys) from the AKFIN data server.
}
