% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_event_fish.R
\name{join_event_fish}
\alias{join_event_fish}
\title{Get EMA fish specimen data with event information}
\usage{
join_event_fish(start_year = 2002, end_year = 3000, tsn = NA, gear = "CAN")
}
\description{
This function pulls specimem level data with associated event information from
ecosystem survey data (from EMA surveys) from the AKFIN data server.
It includes parameters to control start and end year, gear type, and TSN (species code).

You may want to catch weight the specimen data in which case you should use the join_event_catch
function to get the associated catch information. You then join on station ID, event code, and
event code.
}
