% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_fxn.R
\name{get_ema_catch}
\alias{get_ema_catch}
\title{Get ema data}
\usage{
get_ema_catch(start_year, end_year, tsn, gear)
}
\description{
This function pulls ecosystem survey data (from EMA surveys) from the AKFIN data server
It includes parameters to control start and end year, gear type, and TSN (species code)
}
