% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ema_event_parameters.R
\name{get_ema_event_parameters}
\alias{get_ema_event_parameters}
\title{Get ema event table}
\usage{
get_ema_event_parameters(force_download = FALSE)
}
\arguments{
\item{force_download}{Bypass cache and force download}
}
\value{
a dataframe with all calculated event parameters
}
\description{
This function pulls ecosystem survey data (from EMA surveys) from the AKFIN data server
}
