% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_event_catch.R
\name{join_event_catch}
\alias{join_event_catch}
\title{Get EMA catch and event data}
\usage{
join_event_catch(start_year = 1990, end_year = 3000, tsn = NA, gear = NA)
}
\description{
This function pulls ecosystem survey data (from EMA surveys) from the AKFIN data server
It includes parameters to control start and end year, gear type, and TSN (species code)
}
